container/skills-catalog/local/app-walkthrough/requirements.txt
playwright>=1.40.0
pillow>=10.0.0
anthropic>=0.34.0
pyyaml>=6.0
pytest>=7.0
